Move test to their respective package
